backend/requirements.txt
# FastAPI and web server
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.3.6

# LlamaIndex - pinned versions for faster builds
llama-index-core==0.12.36
llama-index-vector-stores-postgres==0.3.3
llama-index-embeddings-huggingface==0.4.0
llama-index-llms-anthropic==0.7.0
llama-index-readers-file==0.4.0

# Anthropic SDK
anthropic>=0.50.0

# Document parsing
pypdf==5.1.0
python-docx==1.1.0
pymupdf==1.23.8

# Additional utilities
python-dotenv==1.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-mock==3.12.0
httpx==0.26.0
